fix: strdup
