+++ include/kernel/string.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include_next "string.hpp"
+
+char* strdup(const char* s);
