+++ include/lib/string.hpp
@@ -14,10 +14,6 @@ const char* strchr(const char* s, char c);
 long strtol(const char* s, const char** endptr = nullptr, int base = 0,
             int n = 0);
 
-#if __has_include("mm/mm.hpp")
-char* strdup(const char* s);
-#endif
-
 class string_view {
   const char* buf_;
   long size_;
